+++ limen/experiment/__init__.py
@@ -1,4 +1,6 @@
 from limen.experiment.experiment_core import UniversalExperimentLoop
+from limen.experiment.manifest_core import CalibrationBuilder
+from limen.experiment.manifest_core import CalibrationConfig
 from limen.experiment.manifest_core import Manifest
 from limen.experiment.param_search import GridStrategy
 from limen.experiment.param_search import RandomStrategy
@@ -9,6 +11,8 @@ from limen.experiment.trainer import Trainer
 
 __all__ = [
     'STRATEGY_REGISTRY',
+    'CalibrationBuilder',
+    'CalibrationConfig',
     'GridStrategy',
     'Manifest',
     'RandomStrategy',